requirements.txt
matplotlib
numpy
pandas
keras
keras-gpu
scikit_learn
opencv-python
Pillow
eventlet
flask
python-socketio
